+++ test/accountTransaction.test.ts
@@ -4,22 +4,34 @@ import {
     AccountTransactionHeader,
     AccountTransactionSignature,
     AccountTransactionType,
+    AttributeKey,
+    IdentityInput,
+    IndexedCredentialDeploymentInfo,
     SimpleTransferPayload,
     SimpleTransferWithMemoPayload,
+    UpdateCredentialsPayload,
+    VerifyKey,
 } from '../src/types';
 import * as ed from 'noble-ed25519';
-import { getAccountTransactionSignDigest } from '../src/serialization';
-import { getNodeClient } from './testHelpers';
+import {
+    getAccountTransactionSignDigest,
+    getCredentialForExistingAccountSignDigest,
+} from '../src/serialization';
+import { getIdentityInput, getNodeClient } from './testHelpers';
 import { AccountAddress } from '../src/types/accountAddress';
 import { GtuAmount } from '../src/types/gtuAmount';
 import { TransactionExpiry } from '../src/types/transactionExpiry';
 import { Memo } from '../src/types/Memo';
+import {
+    buildSignedCredentialForExistingAccount,
+    createUnsignedCredentialForExistingAccount,
+} from '../src/credentialDeploymentTransactions';
 
 const client = getNodeClient();
-const senderAccountAddress =
-    '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
 
 test('send simple transfer signed with wrong private key is accepted', async () => {
+    const senderAccountAddress =
+        '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
     const simpleTransferPayload: SimpleTransferPayload = {
         amount: new GtuAmount(100n),
         toAddress: new AccountAddress(
@@ -68,6 +80,8 @@ test('send simple transfer signed with wrong private key is accepted', async () 
 });
 
 test('send simple transfer with memo signed with wrong private key is accepted', async () => {
+    const senderAccountAddress =
+        '4ZJBYQbVp3zVZyjCXfZAAYBVkJMyVj8UKUNj9ox5YqTCBdBq2M';
     const payload: SimpleTransferWithMemoPayload = {
         amount: new GtuAmount(100n),
         toAddress: new AccountAddress(
@@ -111,6 +125,114 @@ test('send simple transfer with memo signed with wrong private key is accepted',
 
     const result = await client.sendAccountTransaction(
         simpleTransferAccountTransaction,
+        signatures
+    );
+    expect(result).toBeTruthy();
+});
+
+test('update credential is accepted', async () => {
+    const identityInput: IdentityInput = getIdentityInput();
+
+    const lastFinalizedBlockHash = (await client.getConsensusStatus())
+        .lastFinalizedBlock;
+    const cryptographicParameters = await client.getCryptographicParameters(
+        lastFinalizedBlockHash
+    );
+    if (!cryptographicParameters) {
+        throw new Error('Missing global');
+    }
+
+    const publicKeys: VerifyKey[] = [
+        {
+            schemeId: 'Ed25519',
+            verifyKey:
+                'b85c8a998d9d2eca2a1f9e8fc41bff045ccb28f5ba3f6462a41f34801af8e898',
+        },
+    ];
+
+    const threshold = 1;
+    const credentialIndex = 1;
+
+    // The attributes to reveal on the chain.
+    const revealedAttributes: AttributeKey[] = ['firstName', 'nationality'];
+
+    const senderAccountAddress =
+        '4mqk9WnmunRAtT8dadH2Xt5ZapnxSnoDD22DRe436qwL8tdbam';
+    const address = new AccountAddress(senderAccountAddress);
+
+    const credentialToDeploy = createUnsignedCredentialForExistingAccount(
+        identityInput,
+        cryptographicParameters.value,
+        threshold,
+        publicKeys,
+        credentialIndex,
+        revealedAttributes,
+        address
+    );
+
+    const credentialDigestToSign = getCredentialForExistingAccountSignDigest(
+        credentialToDeploy.unsignedCdi,
+        address
+    );
+    const credentialSigningKey =
+        '6563f93a822c7d9b8a107558438c1e981bdf9921f2f0bec3efbf96307f4180ad';
+    const credentialSignature = Buffer.from(
+        await ed.sign(credentialDigestToSign, credentialSigningKey)
+    ).toString('hex');
+
+    const signedCredentialToDeploy = buildSignedCredentialForExistingAccount(
+        credentialToDeploy.unsignedCdi,
+        [credentialSignature]
+    );
+
+    const addCredential: IndexedCredentialDeploymentInfo = {
+        cdi: signedCredentialToDeploy,
+        index: 1,
+    };
+
+    const updateCredentialsPayload: UpdateCredentialsPayload = {
+        newCredentials: [addCredential],
+        currentNumberOfCredentials: BigInt(2),
+        threshold: 1,
+        removeCredentialIds: [
+            'ad88734c858172e24cf2d16954da7d17af1b24de8e6b7631aef991556d05065f11ce994108f8f90c8efff9fdd55f1baa',
+        ],
+    };
+
+    const nextAccountNonce = await client.getNextAccountNonce(address);
+    if (!nextAccountNonce) {
+        throw new Error('Nonce not found');
+    }
+
+    const header: AccountTransactionHeader = {
+        expiry: new TransactionExpiry(new Date(Date.now() + 3600000)),
+        nonce: nextAccountNonce.nonce,
+        sender: address,
+    };
+
+    const updateCredentialsAccountTransaction: AccountTransaction = {
+        header: header,
+        payload: updateCredentialsPayload,
+        type: AccountTransactionType.UpdateCredentials,
+    };
+
+    const privateKey =
+        '103dad2edcb1e245204db1f7845358d7d5ae1508e601732e46ebcbde9fb7667f';
+
+    const hashToSign = getAccountTransactionSignDigest(
+        updateCredentialsAccountTransaction
+    );
+    const signature = Buffer.from(
+        await ed.sign(hashToSign, privateKey)
+    ).toString('hex');
+    const signatures: AccountTransactionSignature = {
+        0: {
+            0: signature,
+        },
+    };
+
+    const result = await client.sendAccountTransaction(
+        updateCredentialsAccountTransaction,
         signatures
     );
     expect(result).toBeTruthy();
